refactor: split mainpage into css and js modules
