--- src/Mainpage/Mainpage.ts
@@ -1,8 +1,0 @@
-import './Mainpage.less';
-
-import {libHeader, searchInjection} from './modules/typescript/index-header';
-import libSummary from './modules/typescript/summary';
-
-libHeader();
-searchInjection();
-libSummary();
