+++ src/Mainpage_CSS/Mainpage.less
@@ -1,11 +1,11 @@
-@import './modules/less/episodes.less';
-@import './modules/less/fanmadeshow.less';
-@import './modules/less/index-header.less';
-@import './modules/less/home-card-news.less';
-@import './modules/less/navigation.less';
-@import './modules/less/news.less';
-@import './modules/less/signup.less';
-@import './modules/less/summary.less';
+@import './modules/episodes.less';
+@import './modules/fanmadeshow.less';
+@import './modules/index-header.less';
+@import './modules/home-card-news.less';
+@import './modules/navigation.less';
+@import './modules/news.less';
+@import './modules/signup.less';
+@import './modules/summary.less';
 
 // 清除citizen样式
 .page-有兽档案馆_首页 {
@@ -22,7 +22,7 @@
 }
 
 /* index body */
-.page-有兽档案馆_首页,
+.page-有兽档案馆_首页.action-view,
 .page-Template_首页 {
 	// 主页按钮
 	.btn {
