+++ src/Mainpage_JS/LICENSE
@@ -1,0 +1,5 @@
+/**
+ * SPDX-License-Identifier: CC-BY-SA-4.0
+ * _addText: '{{Gadget Header|license=CC-BY-SA-4.0|import=no}}'
+ * @author 顶呱呱的阿杰 
+ */
