+++ src/Mainpage_JS/Mainpage.ts
@@ -1,0 +1,6 @@
+import {libHeader, searchInjection} from './modules/index-header';
+import libSummary from './modules/summary';
+
+libHeader();
+searchInjection();
+libSummary();
