+++ src/Mainpage_JS/modules/util/animateNumber.ts
